src/dashboard_requirements.txt
dash
dash-bootstrap-components
plotly
paho-mqtt
python-dotenv
twilio
requests
